+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## 🐛 Bug Fixes
 
 * No blinking on analysis pages if more than 1000 travels
+* Correctly handle not supported purposes
+* Correctly handle not supported manual modes
+
 
 ## 🔧 Tech